More lint fixes

+++ ui/src/app/utils/runtime-utils.ts
@@ -1,15 +1,15 @@
+import {runtimeApi} from 'app/services/swagger-fetch-clients';
+import {useOnMount} from 'app/utils';
+import {switchCase} from 'app/utils';
+import {
+  LeoRuntimeInitializer,
+} from 'app/utils/leo-runtime-initializer';
 import {
   runtimeStore,
   useStore
 } from 'app/utils/stores';
-import {useOnMount} from 'app/utils';
-import {
-  LeoRuntimeInitializer,
-} from 'app/utils/leo-runtime-initializer';
 import {Runtime, RuntimeStatus} from 'generated/fetch';
-import {runtimeApi} from 'app/services/swagger-fetch-clients';
 import * as fp from 'lodash/fp';
-import {switchCase} from 'app/utils';
 
 import * as React from 'react';
 
@@ -17,44 +17,46 @@ const {useState, useEffect} = React;
 
 export type RuntimeStatusRequest = 'Delete';
 
-export const RuntimeStateRequest= {
-    Delete: 'Delete' as RuntimeStatusRequest
+export const RuntimeStateRequest = {
+  Delete: 'Delete' as RuntimeStatusRequest
 };
 
 // useRuntime hook is a simple hook to populate the runtime store.
 // This is only used by other runtime hooks
 const useRuntime = (currentWorkspaceNamespace) => {
   useOnMount(() => {
-    const getRuntime = async () => {
-      const leoRuntime = await runtimeApi().getRuntime(currentWorkspaceNamespace)
+    const getRuntime = async() => {
+      const leoRuntime = await runtimeApi().getRuntime(currentWorkspaceNamespace);
       runtimeStore.set({
         workspaceNamespace: currentWorkspaceNamespace,
         runtime: leoRuntime
       });
-    }
+    };
     getRuntime();
   });
-}
+};
 
 // useRuntimeState hook can be used to change the state of the runtime
 // Only 'Delete' is supported at the moment
 export const useRuntimeState = (currentWorkspaceNamespace): [RuntimeStatus, Function]  => {
   const [requestedRuntimeState, setRuntimeState] = useState<RuntimeStatusRequest>();
   const {runtime} = useStore(runtimeStore);
-  useRuntime(currentWorkspaceNamespace)
+  useRuntime(currentWorkspaceNamespace);
 
   useEffect(() => {
     // Additional state changes can be put here
-    !!requestedRuntimeState && switchCase(requestedRuntimeState,
-      [RuntimeStateRequest.Delete, async () => {
-        await runtimeApi().deleteRuntime(currentWorkspaceNamespace);
-        await LeoRuntimeInitializer.initialize({workspaceNamespace: currentWorkspaceNamespace, maxCreateCount: 0});
-      }])
+    if (!!requestedRuntimeState) {
+      switchCase(requestedRuntimeState,
+        [RuntimeStateRequest.Delete, async() => {
+          await runtimeApi().deleteRuntime(currentWorkspaceNamespace);
+          await LeoRuntimeInitializer.initialize({workspaceNamespace: currentWorkspaceNamespace, maxCreateCount: 0});
+        }]);
+    }
 
-  }, [requestedRuntimeState])
+  }, [requestedRuntimeState]);
 
-  return [runtime.status, setRuntimeState]
-}
+  return [runtime.status, setRuntimeState];
+};
 
 // useCustomRuntime Hook can request a new runtime config
 // The LeoRuntimeInitializer could potentially be rolled into this code to completely manage
@@ -62,10 +64,10 @@ export const useRuntimeState = (currentWorkspaceNamespace): [RuntimeStatus, Func
 export const useCustomRuntime = (currentWorkspaceNamespace): [Runtime, Function] => {
   const {runtime, workspaceNamespace} = useStore(runtimeStore);
   const [requestedRuntime, setRequestedRuntime] = useState<Runtime>();
-  useRuntime(currentWorkspaceNamespace)
+  useRuntime(currentWorkspaceNamespace);
 
   useEffect(() => {
-    const runAction = async () => {
+    const runAction = async() => {
       await runtimeApi().deleteRuntime(currentWorkspaceNamespace);
       await LeoRuntimeInitializer.initialize({
         workspaceNamespace,
@@ -79,11 +81,13 @@ export const useCustomRuntime = (currentWorkspaceNamespace): [Runtime, Function]
           runtime: currentRuntime
         });
       }
-    }
+    };
 
-    requestedRuntime !== undefined && !fp.equals(requestedRuntime, runtime) && runAction();
+    if (requestedRuntime !== undefined && !fp.equals(requestedRuntime, runtime)) {
+      runAction();
+    }
   }, [requestedRuntime]);
 
 
-  return [runtime, setRequestedRuntime]
-}
+  return [runtime, setRequestedRuntime];
+};
